Restore "list enqueue" function of sending queue
In production we found the "enqueue" stage of the pipeline was taking a
non-trivial amount of time (up to 12ms per batch). One possible reason
is that each individual notification was being enqueued with a separate
GenServer call, creating many opportunities for contention with the pop
calls from the sending workers.

The previous "stack" version of the sending queue had a function to
enqueue a list of notifications in a single call. This restores it.