Increase number of sending workers to 50
Based on production metrics, this will be needed to send emails for
large alerts in a timely fashion, as the average time for a single SES
call is ~150ms. With 50 workers, this gives a send rate of 333/s.

To avoid excessive idle load on the sending queue process, which could
slow down the "enqueue" calls that make the work available, this also
increases the amount of time a worker will "back off" if the queue is
found to be empty. Originally, with 2 workers and a delay of 100ms, the
queue would get 20 calls/s at idle. Now with 50 workers and a 1s delay,
the queue will get 50 calls/s.