Fix logging of courtesy emails
The version of the alert list passed to the courtesy email code did not
include the tracking fields, so the `NotificationSender` would later be
unable to log the sending correctly. This fixes that, while ensuring we
can filter courtesy emails out of the stats if needed, by giving them a
distinct "duration type". (courtesy emails ignore the alert processor's
current duration type and are sent for all alerts in the feed)