Fix safeguard for secrets deletion on deploy
From the documentation for the `-e` option of `set`:

> When this option is on, if a simple command fails for any of the
> reasons listed in `Consequences of Shell Errors` or returns an exit
> status value >0, and is not part of the compound list following a
> **while**, **until**, or **if** keyword, and is not a part of an AND
> or OR list, and is not a pipeline preceded by the **!** reserved word,
> then the shell shall immediately exit.

The command that checks for secrets deletion happens to be a pipeline
preceded by the `!` reserved word, thus fulfilling an oddly specific
condition under which `set -e` has no effect. This can be worked around
by wrapping the command in a subshell, but we can also make this check a
bit more "user-friendly" with a conditional.