Adjust timing of status checks in deploy script
Since we've now reduced the deregistration delay on the load balancer
from 300 seconds to 5 seconds, we don't expect deploys to take quite as
long. We can check the status of AWS operations more frequently; this
also puts a lower limit on how long we'll wait before assuming something
went wrong.