Fix password resets with non-lowercase emails
Although we allowed any casing of a user's email to be entered in the
password reset form, the resulting link would only work if the email was
entered in all lowercase, and would otherwise give a confusing 404 error
page when the new-password form was submitted.

Since we downcase user emails whenever they enter the database via user
create or update, it makes sense that we'd also want to downcase the
input whenever we look up a user by email, so this implements the change
within `User.for_email`.