Fix crash when all service to a stop is suspended
We noticed this because all service is currently suspended on `Boat-F4`,
causing its stops to no longer be considered "on" the route in the API,
but it would have already been affecting users since (at least) the
closure of Winchester Center on the Lowell line.

This is a quick fix — it's theoretically possible for us to display the
correct stop name in these cases instead of "Unknown Stop", since
affected stops still exist in the API in general, but under the current
ServiceInfoCache model stops only exist if they're "on" a route.

This also adjusts some tests to otherwise avoid using `Boat-F4`, and
tweaks one test to avoid running into the temporary shuttle-related
change where stations on the Fitchburg line now incorporate nearby bus
stops as child stops.