fix: subscriptions page crash with removed route
If a user had any subscriptions involving a route that didn't exist
anymore, the subscriptions page would crash on render, preventing any
further managing of subscriptions. This also blocked the test suite from
passing, as the recently eliminated bus route 57A was used in a few test
cases.

This fixes the immediate issue by displaying "Unknown Route" as the
route name. For now attempting to edit such a subscription will still
cause a crash, but it can be deleted, and will otherwise not cause any
harm.