fix: alert processing crash in `get_route`
In c4914b5 we removed a seemingly unused clause of `get_route` that
allowed passing `nil` as the route ID. Though this was not covered by
any tests, it turned out to be used when an uncommon set of conditions
occurred, and this caused processing of the "oldest" group of alerts to
always crash in production.

This adds two tests that both fail in the presence of this bug: one for
`ServiceInfoCache` that explicitly tests passing `nil` as the route ID,
and one for `InformedEntityFilter` replicating the conditions that cause
`nil` to be passed as the route ID.

This also fixes an incorrect pattern in `InformedEntityFilter` that
never would have matched, because it was matching a subscription `type`
(which only has atom values) against a string.