refactor: enable running on multiple instances
This is a narrowly-targeted change to address our lack of production
redundancy, and allow the app to be deployed without taking it offline
for the duration. In particular, this does not enable alert processing
to scale horizontally.

There are a few moving parts:

* `AlertProcessor.Lock` is introduced, a database-backed lock that can
  only be acquired by one instance at a time.

* `Model.Metadata` is introduced, a method of storing generic key-value
  data in the database.

* `AlertWorker` is rewritten to use `Lock` to ensure no other instances
  are processing alerts at the same time, and `Metadata` to determine
  when each duration type was last processed, even if it was processed
  by a different instance.

* `deploy.sh` is modified to just update the service with the new task
  definition, instead of taking it down, updating it, and bringing it
  back up. Most of this is borrowed from the API dev-blue deploy script.

* `prod.vm.args` is created for the Distillery release, allowing us to
  set the `sname` option. This causes a "hostname" like `ip-x-x-x-x` to
  be generated at runtime, which will then appear in the "node" part of
  log lines and allow us to differentiate between instances.

Since alert processing speed has improved quite a bit since the numbers
were originally set, this also adjusts the interval for "oldest" alerts,
which are now processed every 10 minutes instead of every 60. Typically
this group only takes ~10 seconds to process on production.