config: use default database pool size in prod
We previously increased the pool size (see #860) to work around this
issue: https://github.com/elixir-ecto/db_connection/issues/127

As of recently (see #1129) we are on a version of OTP greater than the
one that fixes the issue. Also, we are now starting to run the app on
more than one instance at a time, and the large number of connections
opened by the pool was causing additional instances to crash on startup
with the error `FATAL 53300 (too_many_connections)`.