feat: handle email bounces and complaints
We had partially set up an endpoint intended to handle these messages,
but it didn't do anything other than log them, didn't verify signatures,
and was not subscribed to the SNS topic anyway.

This revamps the endpoint to correctly respond to complaints and
permanent bounces by setting the user's communication mode to `none`,
similar to how SMS opt-out works. This prevents sending any further
notifications until the user next updates their account. We also set an
`email_rejection_status` field, both to indicate the reason the mode is
`none` (since there are now multiple reasons it could be), and to enable
better UI for this feature later.

Currently when the mode is `none` it will appear to be "Email" on the
account page. This was also an issue with SMS opt-out, so it can be
addressed separately, and shipping this change sooner is better so we
can start reducing our bounce/complaint rates.