feat: explain why alerts are disabled
When a user's communication mode was `none`, unless they opted out via
SMS and were currently inside the freeze period, the UI did not explain
this state or alert the user to it.

With these changes, if a user's communication mode is `none`:

* a flash appears on their Settings page explaining how this happened
  and what they can do to resolve it (for all circumstances, not just
  when inside the SMS freeze period)

* a flash appears on their Subscriptions page, which is the "home page"
  when signed in, explaining that alerts are disabled and to access the
  Settings page to resolve this

In any case where the SMS option is disabled, there is also a flash
explaining why, even if the communication mode is not `none`.