ci: enable manual cache-busting without commit
`CACHE_UUID` is a random UUID, stored as a repo secret because it's a
convenient way to store workflow-accessible data that isn't committed
to the repo. If we need to bust the cache for any reason, we can change
this secret to a different random UUID.

Separately, we can remove `runner.os` from cache keys, since we don't
currently run builds across multiple operating systems (the value of
this is just e.g. `Linux` and not the specific distribution or version
of Linux).