fix: missing favicon
This occurred because the `<link rel="icon">` in the HTML used
the "digested" filename in production, but following this link would
give a 404, since the site was only configured to serve the un-digested
`favicon.ico`. Depending on whether the browser "preferred" the `<link>`
or the file at `/favicon.ico`, it sometimes wouldn't show the icon.

Using `only_matching` instead of `only` in `Plug.Static` allows serving
digested files at the root, since strings specified here only need to
be a prefix of the first path component instead of matching it exactly.