fix: insecure use of `String.to_atom/1`
In both of these cases, the atom should already exist (unless we made a
mistake in code or someone is submitting rogue HTTP requests), so we can
use `String.to_existing_atom/1`.

This removes the "feature" of `IconViewHelper.icon_for_route/2` that the
mode and the route could both be passed as either an atom or a string.
In practice this was never used, since the first argument always comes
from `Subscription.type` which is an atom, and the second always from
`Subscription.route` which is a string (unless it's a list, smuggled in
by `TripController.expand_multiroute_greenline_subscription`).