fix: 500 error when auth token expired
Guardian 0.14.x had a default set of error handlers it would fall back
to if you didn't handle a particular type of error. This is no longer
the case in Guardian 1.x due to the error handler API changing to use
different clauses of a single callback, instead of multiple callbacks.
This means we need to include a fallback clause ourselves.