fix: apply rate limiting to all non-GET actions
This was a security consideration raised when a vulnerability scanning
tool briefly "attacked" the site with mass requests. This expands rate
limiting from just the sign-in form to any action that submits data (is
not a GET request).

Since the new solution is global instead of per-action, for now we have
no use for `hammer_plug` anymore, since it does not provide conveniences
for global rate limiting based on the method of the request.