feat(admin): filter inactive users out of reports
Following a vulnerability scan that created a massive number of fake
accounts, complete with fake subscriptions, the numbers in our reports
were a bit skewed. We can compensate for this by ignoring accounts
where alerts are disabled, which happens when either their email
address bounces (which we saw with these accounts, since they made up
non-existent addresses) or they opt out via SMS. In cases where this
occurs "naturally", it should still be okay to not include those
accounts in the reports, since we are primarily interested in stats on
current users of the app.