fix: JS error in subscription creation
When creating a bus "leg" for a subscription, there are no origin and
destination selectors, but the JS code that applied client-side errors
to these selectors was still running, and crashing when it didn't find
them. This had no user-facing impact, since the JS runs on form submit,
which (due to the error) was not prevented and immediately moved users
to the next page anyway.