fix: reversed route icon order in stop selectors
Changes to attribute ordering due to the Phoenix upgrade in fe6c785c
introduced a bug where route icons in stop selectors would display in
the opposite of the intended order. This was especially noticeable with
the Green Line routes, which were ordered "E D C B". Since the reversed
ordering is a consequence of using `float: right`, we can un-reverse it
in the same place we apply this style.