ci: use `npm ci` instead of `npm install`
This installs dependencies exactly as specified in the lockfile, and
always exits with an error in cases where `npm install` would update
`package.json` and/or the lockfile (e.g. if the lockfile did not match
the version constraints in `package.json`). Also, it's faster. ⚡️

Previously we didn't use this because our NodeJS version was too old for
the included version of NPM to support this command.

Ref: https://docs.npmjs.com/cli/v6/commands/npm-ci