chore: track notification batches
We want to track some operational performance metrics in Splunk. One thing
that we want to track is the total time taken to send all notifications for
an alert. However, since alerts can persist over longer periods of time,
tracking just the `alert_id` is insufficient. Thus, with this PR we want to
add a concept of a `batch_id`. This will allow us to see how long it takes
for us to send all notifications for an alert _in a given batch_ rather
than over the entire lifetime of the alert.