feat: Remove local auth support (#1350)
* feat: Remove all `keycloak_auth?` branching

Now that all of our environments have switched to Keycloak for auth, we
no longer need to support the local auth option.

* fix: Skip redirecting to OIDC logout on failed login

In this case, we know that they were never actually logged in.

* feat: Remove unused authentication_source env var

Don't run OIDC worker in test.

* feat: Remove local password reset functionality

Including controller, model, and related templates.

* feat: Drop the password_resets database table

* feat: Remove Admins controller and routes

User roles are now controlled in Keycloak.

* feat: Remove user passwords

This is now handled by Keycloak.

* feat: Remove user email validation

This is now handled by Keycloak. The email address is actually verified
before the user can make it to T-Alerts.