fix: Log as warning for sending failures that get retried
These 400 failures typically represent throttling responses from AWS. We
are able to retry sending, which generally results in success. Given
this, a warning log message is sufficient. This prevents us from
triggering an alert due to an expected situation when we are sending a
high volume of notifications in a short time period.